server/requirements.txt
pypdf==5.1.0
chromadb==0.5.20
pytest==8.3.3
boto3==1.35.65
langchain==0.3.7
langchain-chroma==0.1.4
langchain-ollama==0.2.0
langchain-community==0.3.7
Django==5.1.3
django
Chroma==0.2.0
ollama==0.3.3
django-browser-reload==1.15.0
daphne==4.1.2
django-eventstream==5.3.1
djangorestframework==3.14.0
django-cors-headers==3.13.0
